moved xmlreader to separate file and created subfolder for import-export
